Move AppGraph and AppScope to app/di package
Consolidates all DI configuration in app/di/ package.
