+++ androidApp/src/main/kotlin/space/be1ski/vibits/android/MainActivity.kt
@@ -8,7 +8,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import space.be1ski.vibits.shared.app.AppRoot
-import space.be1ski.vibits.shared.di.AppGraph
+import space.be1ski.vibits.shared.app.di.AppGraph
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
