+++ desktopApp/src/desktopMain/kotlin/space/be1ski/vibits/desktop/DesktopMain.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
 import space.be1ski.vibits.shared.app.AppRoot
-import space.be1ski.vibits.shared.di.AppGraph
+import space.be1ski.vibits.shared.app.di.AppGraph
 
 fun main() =
   application {
