+++ shared/src/iosMain/kotlin/space/be1ski/vibits/shared/presentation/MainViewController.kt
@@ -2,8 +2,8 @@ package space.be1ski.vibits.shared.presentation
 
 import androidx.compose.ui.window.ComposeUIViewController
 import platform.UIKit.UIViewController
+import space.be1ski.vibits.shared.app.di.AppGraph
 import space.be1ski.vibits.shared.app.view.AppRoot
-import space.be1ski.vibits.shared.di.AppGraph
 
 /**
  * Entry point for embedding Compose UI into an iOS host.
