+++ webApp/src/wasmJsMain/kotlin/space/be1ski/vibits/web/Main.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import kotlinx.browser.document
 import space.be1ski.vibits.shared.app.AppRoot
-import space.be1ski.vibits.shared.di.AppGraph
+import space.be1ski.vibits.shared.app.di.AppGraph
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
